+++ src/components/accordion/Accordion.tsx
@@ -3,7 +3,6 @@ import { css } from '@emotion/react';
 import type {
   ReactElement,
   ReactNode,
-  ReactFragment,
   MouseEvent as ReactMouseEvent,
 } from 'react';
 import { useCallback } from 'react';
@@ -14,7 +13,6 @@ export interface AccordionProps {
   children?:
     | Array<ReactElement<AccordionItemProps>>
     | ReactElement<AccordionItemProps>
-    | ReactFragment
     | boolean
     | null;
 }
@@ -67,7 +65,7 @@ Accordion.Item = function AccordionItem(props: AccordionItemProps) {
   const { item, utils } = useAccordionContext(title, defaultOpened);
 
   const onClickHandle = useCallback(
-    (event: ReactMouseEvent<HTMLButtonElement>) => {
+    (event: ReactMouseEvent<HTMLDivElement>) => {
       if (event.shiftKey) {
         utils.clear();
       } else {
@@ -85,9 +83,9 @@ Accordion.Item = function AccordionItem(props: AccordionItemProps) {
         flexDirection: 'column',
       }}
     >
-      <button
+      <div
         onClick={onClickHandle}
-        type="button"
+        role="button"
         style={{
           display: 'flex',
           alignItems: 'center',
@@ -106,7 +104,7 @@ Accordion.Item = function AccordionItem(props: AccordionItemProps) {
           {title}
         </div>
         {toolbar}
-      </button>
+      </div>
       <div
         style={{
           display: item?.isOpen ? 'flex' : 'none',