+++ src/components/root-layout/root_layout_context.ts
@@ -1,0 +1,12 @@
+import { createContext, useContext } from 'react';
+
+const defaultPortalContext =
+  typeof document === 'undefined' ? null : document.body;
+
+export const rootLayoutContext = createContext<HTMLElement | null>(
+  defaultPortalContext,
+);
+
+export function useRootLayoutContext() {
+  return useContext(rootLayoutContext);
+}