+++ src/components/table/tableContext.ts
@@ -1,7 +1,0 @@
-import { createContext, useContext } from 'react';
-
-export const TableContext = createContext({ color: '' });
-export function useTableContext() {
-  const context = useContext(TableContext);
-  return context;
-}