+++ src/pages/demo/App.tsx
@@ -1,3 +1,4 @@
+import { FifoLogger } from 'fifo-logger';
 import { KbsProvider } from 'react-kbs';
 
 import { AppStateProvider } from '../../app-data';
@@ -11,8 +12,6 @@ import MainLayout from './MainLayout';
 
 import '@blueprintjs/core/lib/css/blueprint.css';
 import '@blueprintjs/icons/lib/css/blueprint-icons.css';
-
-import { FifoLogger } from 'fifo-logger';
 
 const fifoLogger = new FifoLogger({ level: 'debug' });